[feat] Lotto 도메인
